more modularization (including some cleanup)
