+++ static/scripts/dataprivacy/registration.js
@@ -1,4 +1,5 @@
 import '../helpers/inputLinking';
+import './pin';
 
 window.addEventListener('DOMContentLoaded', ()=>{
     // show steppers depending on age of student
@@ -32,43 +33,7 @@ window.addEventListener('DOMContentLoaded', ()=>{
             }
         });
     }
-    let pinSent;
-    // if email for pin registration is changed, reset pin-sent status
-    $('form.registration-form.student input[name$="email"]:last').on("change", ()=> {
-        pinSent = false;
-    });
-    
-    $('.form section[data-feature="pin"]').on("showSection", (event) => {
-        if(pinSent) {
-            // send pin of value is something else than no
-        } else {
-            sendPin(true);
-        }
-    });
-    
-    $('#resend-pin').on("click", e => {
-        e.preventDefault();
-        sendPin(true);
-        $(".pin-input .digit").val("");
-    });
-    
-    function sendPin(sendConfirm) {
-        let usermail = $("input[name$='email']:last").val();
-        let byRole = window.location.pathname.split("/by")[1].split("/")[0].replace("/","");
-        
-        $.ajax({
-            url: "/registration/pincreation",
-            method: "POST",
-            data: {"email": usermail, "byRole": byRole}
-        }).done(success => {
-            if(sendConfirm) {
-                $.showNotification(`Eine PIN wurde erfolgreich an ${usermail} versendet.`, "success", 15000);
-            }
-            pinSent = true;
-        }).fail(function(err){
-            $.showNotification(`Fehler bei der PIN-Erstellung! Bitte versuche es mit 'Code erneut zusenden' und prüfe deine E-Mail-Adresse (${usermail}).`, "danger", 7000);
-        });
-    }
+
 
     let parentMailInput = document.querySelector('input[name="parent-email"]');
     let studentMailInput = document.querySelector('input[name="student-email"]');
@@ -108,6 +73,8 @@ function goBack(event){
     window.history.back();
 }
 
+
+// GENERATE START PASSWORD
 window.addEventListener('load', ()=>{
     if(document.querySelector('.form .student-password')) {
         // generate password if password field present
