--- views/firstLogin/sections/consent_parent.hbs
@@ -1,6 +1,0 @@
-<section data-panel="section-{{section-nr}}" {{#if submit-page}}class="submit-page"{{/if}}>
-  <h4>Einwilligungserklärung</h4>
-  <p>Die folgenden Daten müssen von einer erziehungsberechtigten Person ausgefüllt werden!</p>
-  <p>Ich, bin erziehungsberechtigt für <b>{{@root.currentUser.firstName}} {{@root.currentUser.lastName}}</b> und erkläre Folgendes:</p>
-    {{#embed "registration/forms/einwilligungserklaerung"}}{{/embed}}
-</section>
