+++ views/firstLogin/sections/pin.hbs
@@ -1,4 +1,4 @@
-<section data-panel="section-{{section-nr}}" {{#if submit-page}}class="submit-page"{{/if}} data-feature="pin">
+<section data-panel="section-{{section-nr}}" class="submit-page" data-feature="pin">
   <h4>Fast geschafft!</h4>
   <p>Wir haben soeben eine E-Mail an Ihre angegebene Adresse gesendet. Bitte geben Sie den darin enthaltenen Code hier ein, um die Registrierung abzuschließen. Sie haben keine E-Mail erhalten? <button type="button" class="btn-link" id="resend-pin">Code erneut senden</button></p>
   <div id="pinverification">
